rename to logging_ to avoid shadowing default name
